Updates for Node 8.7 and Meteor 1.6. Replaced yarn with npm.
Meteor 1.6 upgraded it's version of Node.js to 8.7, which means npm 5.4.2. Npm now installs very fast, so there isn't much of a reason to run yarn and maintain the additional complexity with it, so we are reverted from yarn back to npm.

This commit also comes an entirely new builddeploy file that is updated for mobile build support and contains a few additional features.